requirements/fastapi.txt
# FastAPI Requirements for RAG System API
# Multi-user RAG system for academic papers with Gemini and ArXiv integration

# Include base requirements
#  -r base.txt

# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# File upload and handling
python-multipart==0.0.6
aiofiles==23.2.1

# Request/Response validation
pydantic==2.5.0

# HTTP client for async operations
httpx==0.25.2

# CORS and middleware
starlette==0.27.0

# WebSocket support (for real-time features)
websockets==12.0

# Authentication (JWT tokens and password hashing)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
email-validator==2.1.0

# Background task processing
celery==5.3.4
redis==5.0.1

# Username normalization
python-slugify==8.0.1
Unidecode==1.3.7

# Performance monitoring
psutil==5.9.6

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2  # For testing async endpoints

# Code quality (optional)
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# Additional utilities
python-dateutil>=2.8.0
pytz>=2023.3